Move and test loadAccounts
